requirements.txt
# Core Dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
oauth2client==4.1.3

# AI/ML Services
openai==1.3.0
anthropic==0.8.0

# Video Processing
opencv-python==4.8.1.78
moviepy==1.0.3
pillow==10.1.0

# Audio Processing
pydub==0.25.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Optional - for enhanced features
# elevenlabs==0.2.26  # Text-to-speech
# replicate==0.15.0   # AI image generation
# stability-sdk==0.8.4  # Stable Diffusion
